# rest-api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:9000",
  "basePath": "/",
  "tags": [
    {
      "name": "game-statistics-controller",
      "description": "Game Statistics Controller"
    }
  ],
  "paths": {
    "/api/{id}": {
      "get": {
        "tags": [
          "game-statistics-controller"
        ],
        "summary": "getStatistics",
        "operationId": "getStatisticsUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GameStatisticsRes"
            }
          }
        }
      },
      "put": {
        "tags": [
          "game-statistics-controller"
        ],
        "summary": "updateStatistics",
        "operationId": "updateStatisticsUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "gameStatistics",
            "description": "gameStatistics",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GameStatisticsReq"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "definitions": {
    "GameStatisticsReq": {
      "type": "object",
      "required": [
        "countOfCorrectAnswers",
        "countOfTests",
        "difficulty",
        "endGameDateTime",
        "solvingTime"
      ],
      "properties": {
        "countOfCorrectAnswers": {
          "type": "integer",
          "format": "int32",
          "example": 3,
          "description": "Количество правильных ответов"
        },
        "countOfTests": {
          "type": "integer",
          "format": "int32",
          "example": 5,
          "description": "Количество тестов"
        },
        "difficulty": {
          "type": "string",
          "example": 0,
          "description": "Сложность",
          "enum": [
            "0",
            "1",
            "2"
          ]
        },
        "endGameDateTime": {
          "type": "string",
          "format": "date-time",
          "example": "2020-11-25T19:42:18.298Z",
          "description": "Время окончания игры"
        },
        "solvingTime": {
          "type": "integer",
          "format": "int32",
          "example": 100,
          "description": "Среднее время решения здачи в секундах"
        }
      },
      "title": "GameStatisticsReq"
    },
    "GameStatisticsRes": {
      "type": "object",
      "required": [
        "countOfCorrectAnswers",
        "countOfTests",
        "difficulty",
        "endGameDateTime",
        "solvingTime"
      ],
      "properties": {
        "countOfCorrectAnswers": {
          "type": "integer",
          "format": "int32",
          "example": 3,
          "description": "Количество правильных ответов"
        },
        "countOfTests": {
          "type": "integer",
          "format": "int32",
          "example": 5,
          "description": "Количество тестов"
        },
        "difficulty": {
          "type": "string",
          "example": 0,
          "description": "Сложность",
          "enum": [
            "0",
            "1",
            "2"
          ]
        },
        "endGameDateTime": {
          "type": "string",
          "format": "date-time",
          "example": "2020-11-25T19:42:18.298Z",
          "description": "Время окончания игры"
        },
        "solvingTime": {
          "type": "integer",
          "format": "int32",
          "example": 100,
          "description": "Среднее время решения здачи в секундах"
        }
      },
      "title": "GameStatisticsRes"
    }
  }
}